[JAX Runner] Refactor jax runner to delegate speculative decoding logic [4/N]
